Correct health/exp. bar color, correct database bug
- Corrected bug that wouldn't allow player to save character after creating it and buying an item

+++ namelessGame/src/namelessgame/Threads/HealthBar.java
@@ -92,7 +92,7 @@ public class HealthBar extends Thread {
             // value progress bar %
             // total_hp - 100
             // hpLabelValue - x
-            hpBar.setValue((hpLabelValue * 100) / maxHp);
+            hpBar.setValue((int) (((double) hpLabelValue * 100) / maxHp));
             
             try { Thread.sleep (timePerHealthPoint); } catch (InterruptedException ex) {}
         }
